Use structs to represent ascii art in different states of processing
Stop passing raw strings around, as that's error prone.
